Re-add component for keyboard shortcuts
