+++ docs/getting-started/tipstricks.md
@@ -5,44 +5,44 @@ title: 'Tips & Tricks'
 
 ## Undo/redo
 
-If you ever make a mistake, pressing `Ctrl & z` will undo, and `Ctrl, Shift & z` will redo. This is an undo system that you can rely on; any change can be undone and the UI will walk back in time.
+If you ever make a mistake, pressing <Key mod="cmd" k="z" /> will undo, and pressing <Key mod="cmd shift" k="z" /> will redo. This is an undo system that you can rely on; any change can be undone and the UI will walk back in time.
 
 ## Keyboard shortcuts
 
 A few global shortcuts:
 
-- `Ctrl & 1` will show the budget
-- `Ctrl & 2` will show reports
-- `Ctrl & 3` will show all accounts
-- `Ctrl & o` will close the file and list other available files to open
+- <Key mod="cmd" k="1" /> will show the budget
+- <Key mod="cmd" k="2" /> will show reports
+- <Key mod="cmd" k="3" /> will show all accounts
+- <Key mod="cmd" k="o" /> will close the file and list other available files to open
 
 ### Budget
 
-- `Enter` while editing a budget amount will move to the next category.
-- `Shift & Enter` will move to the previous category.
+- <Key k="enter" /> while editing a budget amount will move to the next category.
+- <Key mod="shift" k="enter" /> will move to the previous category.
 
 ### Transaction Editing
 
-- When editing, `Enter` and `Shift & Enter` will move down and up. If a dropdown is open and you've typed a new value, this will instead save the value and close the dropdown.
-- When editing, `Tab` and `Shift & Tab` will move left and right
-- When adding a new transaction, `Ctrl & Enter` will add it regardless of where you are editing. Pressing `Enter` in either the Payment or Deposit columns will add it as well.
+- When editing, <Key k="enter" /> and <Key mod="shift" k="enter" /> will move down and up. If a dropdown is open and you've typed a new value, this will instead save the value and close the dropdown.
+- When editing, <Key k="tab" /> and <Key mod="shift" k="tab" /> will move left and right
+- When adding a new transaction, <Key mod="cmd" k="enter" /> will add it regardless of where you are editing. Pressing <Key k="enter" /> in either the Payment or Deposit columns will add it as well.
 
 ### Transaction Management
 
-- `f` Show only selected transactions.
-- `d` Delete selected transactions.
-- `a` Set account for selected transations.
-- `p` set payee for selected transactions.
-- `n` Set notes for selected transactions.
-- `c` Set category for selected transactions.
-- `l` Toggle cleared for current transaction.
+- <Key k="f" /> Show only selected transactions.
+- <Key k="d" /> Delete selected transactions.
+- <Key k="a" /> Set account for selected transations.
+- <Key k="p" /> set payee for selected transactions.
+- <Key k="n" /> Set notes for selected transactions.
+- <Key k="c" /> Set category for selected transactions.
+- <Key k="l" /> Toggle cleared for current transaction.
 
 The following require a transaction to have been selected first:
 
-- `j/Down Arrow` Move to the next transaction down.
-- `k/Up Arrow` Move to the next transaction up.
-- `Space` Toggle selection of current transaction.
-- `Shift & Space` Add all transactions between current transaction and most recently selected transaction.
+- <Key k="J" />/<Key arrow="down" /> Move to the next transaction down.
+- <Key k="K" />/<Key arrow="up" /> Move to the next transaction up.
+- <Key k="space" /> Toggle selection of current transaction.
+- <Key mod="shift" k="space" /> Add all transactions between current transaction and most recently selected transaction.
 
 ## View multiple months at once
