+++ docs/install/local.md
@@ -56,7 +56,7 @@ When first accessing Actual, you may be prompted to provide a URL for the server
 
 When we publish a new release, you’ll need to follow these steps to update:
 
-1. Stop the server if it’s running. You can use the keyboard shortcut `Ctrl & c` (even on macOS) to stop the server, or close the terminal window it’s running from.
+1. Stop the server if it’s running. You can use the keyboard shortcut <Key mod="ctrl" k="c" /> (even on macOS) to stop the server, or close the terminal window it’s running from.
 2. Run `git pull` from the directory you cloned the project into. This will download the latest server code.
 3. Run `yarn install` from that same directory. This will download the latest web client code, along with any updated dependencies for the server.
 4. Restart the server by running `yarn start`.
