+++ src/theme/MDXComponents.js
@@ -1,7 +1,8 @@
 import MDXComponents from '@theme-original/MDXComponents';
 import Image from '@site/src/components/Image';
-
+import Key from '../components/Key';
 export default {
   ...MDXComponents,
   img: Image,
+  Key,
 };
